requirements.txt
streamlit==1.28.1
google-generativeai==0.3.0
Pillow==10.1.0
sympy==1.12
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
tensorflow==2.14.0
tensorflow-keras==2.14.0
